refactored InfluxDBv2 data listener
